requirements.txt
pandas
ccxt
pandas_ta
pyarrow
pyyaml
scikit-learn
statsmodels
matplotlib
seaborn
jupyter
mlflow
joblib
fastapi
uvicorn
django
scipy
streamlit
plotly
